Covid Chairs Updated
Finding least covid chairs is now a function. Also added a Hybrid algorithm that mixes best fit and least covid chairs. It selects the chair according to the smallest size container that has the least amout of covid chairs. If there are multiple options it then selects a random one. It's performance is sometimes better than best fit and sometimes worse. Im not sure if random selection is the best way.